requirements.txt
fastapi
uvicorn
pytest
mongomock
httpx
pymongo
python-dotenv
python-jose
responses>=0.24.0
motor==3.3.1